feat: refactor ride details handling by creating getRideDetails utility for improved data extraction and integration in RideDetailsModal and LiveRideTracking components
